ctf: Correctly handle packet sizes > 2^31 bits
We use long's for packet sizes, so we should be able to handle
those, just need to make sure the value does not get truncated
to Integer.MAX_VALUE (or even worse, becomes negative!)

Reviewed-on: https://git.eclipse.org/r/37129
Tested-by: Hudson CI
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>

+++ org.eclipse.tracecompass.ctf.core/src/org/eclipse/tracecompass/ctf/core/event/io/BitBuffer.java
@@ -93,7 +93,7 @@ public final class BitBuffer {
         fBuffer = buf;
         setByteOrder(order);
         resetPosition();
-        fBitCapacity = fBuffer.capacity() * BIT_CHAR;
+        fBitCapacity = (long) fBuffer.capacity() * BIT_CHAR;
     }
 
     private void resetPosition() {
